feat: 현직자 추천 리스트 스켈레톤 추가
