+++ src/widgets/home/index.ts
@@ -2,6 +2,8 @@ export { default as SignupConfetti } from './ui/SignupConfetti';
 export { default as RecruitmentLinksTicker } from './ui/RecruitmentLinksTicker';
 export { TechBlogBanner, default as TechBlogTicker } from './ui/TechBlogTicker';
 export { default as ExpertRecommendations } from './ui/ExpertRecommendations';
+export { default as ExpertRecommendationsServer } from './ui/ExpertRecommendationsServer';
+export { default as ExpertRecommendationsSkeleton } from './ui/ExpertRecommendationsSkeleton';
 export { HomeGuardToast } from './ui/HomeGuardToast';
 export { default as GuideButtons } from './ui/GuideButtons';
 export { default as SearchBar } from './ui/SearchBar';
